requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
billiard==4.2.0
blinker==1.7.0
Brlapi==0.8.5
celery==5.4.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloud-init==24.1.3
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
cryptography==41.0.7
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
distro==1.9.0
distro-info==1.7+build1
Django==5.0.6
django-allauth==0.63.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.1
duplicity==2.1.4
environ==1.0
ephem==4.1.5
fasteners==0.18
httplib2==0.20.4
idna==3.6
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
kombu==5.3.7
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
Mako==1.3.2.dev0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
netaddr==0.8.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.2
olefile==0.46
pandas==2.2.2
paramiko==2.12.0
pexpect==4.9.0
pillow==10.2.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
py4j==0.10.9.7
pycairo==1.25.1
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
PyNaCl==1.5.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
pyspark==3.5.1
python-apt==2.7.7+ubuntu1
python-dateutil==2.9.0.post0
python-debian==0.1.49+ubuntu2
python-memcached==1.58
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
redis==5.0.6
requests==2.31.0
rich==13.7.1
screen-resolution-extra==0.0.0
setuptools==68.1.2
six==1.16.0
sqlparse==0.5.0
systemd-python==235
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
usb-creator==0.3.16
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.13
wheel==0.42.0
xdg==5
xkit==0.0.0